refactor(codegen): dedup the three method-call SV emission matches into one shared emitter
The ExprKind::MethodCall emission match (.trunc/.zext/.sext/.resize/
.reverse/Vec reductions/float conversions) existed in three
near-identical copies: the canonical one in codegen/mod.rs plus one in
each pipeline expression emitter (emit_pipeline_stage_expr_str,
emit_pipeline_expr_str). A fix applied to one copy could silently miss
the siblings — that is exactly how the .resize self-determination
miscompile fixed in mod.rs shipped on in the pipeline emitters until
PR #839.

All three call sites now dispatch to a single emit_method_call_str in
the new codegen/method_call.rs, parameterized over:
- an emit_sub closure for receiver/sub-expression emission (the
  pipeline emitters rewrite stage-reg / Stage.field names);
- a try_reverse_chunked closure for the arch#808 portable .reverse
  lowering (mod.rs and pipeline use different resolution helpers);
- a MethodCallHost enum keeping the intentional per-host emission
  divergences explicit and documented in one place: zext's $unsigned
  wrap (Main only), sext's static width via infer_sv_width_str vs
  $bits(...) (Pipeline), and the float conversions (Main only;
  Pipeline falls through to the generic default as it always has).

Emission-preserving by construction and by measurement:
scripts/refactor_diff.sh origin/main compared 1680 emitted
.sv/.archi/sim-source files across the full regression corpus and
backend_equiv fixtures — zero byte diffs.
