+++ src/codegen/mod.rs
@@ -3,6 +3,7 @@ use crate::diagnostics::CompileWarning;
 use crate::lexer::Span;
 use crate::resolve::{Symbol, SymbolTable};
 use crate::typecheck::enum_width;
+use method_call::MethodCallHost;
 
 // Per-construct submodules. Each contributes `pub(super) fn emit_<name>`
 // to `impl Codegen` and lives in its own file mirroring the layout of
@@ -16,6 +17,7 @@ mod fifo;
 mod fp;
 mod fsm;
 mod linklist;
+mod method_call;
 mod module;
 mod pipeline;
 mod ram;
@@ -6942,214 +6944,18 @@ impl<'a> Codegen<'a> {
                 let b = self.emit_expr_str(base);
                 format!("{b}.{}", field.name)
             }
-            ExprKind::MethodCall(base, method, args) => {
-                let b = self.emit_expr_str(base);
-                match method.name.as_str() {
-                    "trunc" => {
-                        if let Some(width) = args.first() {
-                            let w = self.emit_expr_str(width);
-                            // SV size cast: valid on any expression, including compound ones.
-                            // e.g. (count_r + 1).trunc<8>() → 8'(count_r + 1)
-                            let wp = Self::paren_width(&w);
-                            format!("{wp}'({b})")
-                        } else {
-                            b
-                        }
-                    }
-                    "zext" => {
-                        if let Some(width) = args.first() {
-                            let w = self.emit_expr_str(width);
-                            // $unsigned prevents context-dependent width expansion before the cast
-                            let wp = Self::paren_width(&w);
-                            format!("{wp}'($unsigned({b}))")
-                        } else {
-                            b
-                        }
-                    }
-                    "sext" => {
-                        if let Some(width) = args.first() {
-                            let w = self.emit_expr_str(width);
-                            // Sign-extension: replicate the MSB into the upper
-                            // bits. `.sext()` always treats the receiver's own
-                            // MSB as the sign bit regardless of the receiver's
-                            // declared signedness, so a `signed(...)`/
-                            // `unsigned(...)`/`as T` wrapper around the
-                            // receiver (e.g. `signed(raw).sext<16>()`,
-                            // arch#650's "nested" mitigation example) is inert
-                            // here — same bits either way. Unwrap it: besides
-                            // being redundant, indexing straight into a cast
-                            // result (`$signed(raw)[...]`) is exactly the
-                            // "indexed cast" pattern arch#650 flags as
-                            // Icarus-hostile, and `$bits($signed(raw))` (the
-                            // prior width computation) nests a system-function
-                            // call inside `$bits`, which Icarus also does not
-                            // reliably accept.
-                            let recv = Self::unwrap_reinterpret_cast(base);
-                            let rb = self.emit_expr_str(recv);
-                            let sw = self.infer_sv_width_str(recv);
-                            let swp = Self::paren_width(&sw);
-                            format!("{{{{({w}-{swp}){{{rb}[{swp}-1]}}}}, {rb}}}")
-                        } else {
-                            b
-                        }
-                    }
-                    "resize" => {
-                        if let Some(width) = args.first() {
-                            let w = self.emit_expr_str(width);
-                            // Direction-agnostic resize: pads or truncates, preserving
-                            // signedness. SV's `N'(expr)` size cast inherits the
-                            // signedness of `expr` and — critically — forwards
-                            // context-determined evaluation through arithmetic
-                            // operators inside it. Earlier emission used
-                            // `N'($signed(expr))` / `N'($unsigned(expr))`, but
-                            // `$signed`/`$unsigned` evaluate their argument in
-                            // self-determined context (LRM §11.6.1, §20.5), which
-                            // truncates a multiply like `a * b` to operand width
-                            // BEFORE the outer cast widens — silently losing the
-                            // upper bits of any product. Dropping the wrapper lets
-                            // `N'(a * b)` widen both operands to N before the
-                            // multiply. For non-arithmetic `expr` (idents, slices),
-                            // the cast still preserves signedness from the
-                            // underlying declaration, so no behaviour changes.
-                            let wp = Self::paren_width(&w);
-                            format!("{wp}'({b})")
-                        } else {
-                            b
-                        }
-                    }
-                    // as_clock removed — use `as Clock<Domain>` cast syntax // identity — 1-bit logic used as clock
-                    "reverse" => {
-                        if let Some(chunk) = args.first() {
-                            // arch#808: Icarus rejects `{<<N{x}}` in every
-                            // context; prefer the portable chunked-concat
-                            // lowering whenever the receiver width resolves
-                            // (typecheck guarantees const chunk + width, so
-                            // the streaming fallback should be unreachable
-                            // for checked input — kept for safety).
-                            if let Some(s) = self.try_emit_reverse_chunked(base, chunk) {
-                                s
-                            } else {
-                                let c = self.emit_expr_str(chunk);
-                                format!("{{<<{c}{{{b}}}}}")
-                            }
-                        } else {
-                            b
-                        }
-                    }
-                    "any" | "all" | "count" | "contains" | "reduce_or" | "reduce_and"
-                    | "reduce_xor" | "find_first" => self.emit_vec_method(&b, base, method, args),
-                    // Float conversions → emitted helper functions.
-                    "to_fp32" => {
-                        self.fp_helpers_used.set(true);
-                        match self.expr_float_fmt(base) {
-                            Some("bf16") => format!("arch_bf16_to_f32({b})"),
-                            Some("e4m3") => format!("arch_e4m3_to_f32({b})"),
-                            Some("e5m2") => format!("arch_e5m2_to_f32({b})"),
-                            Some("e2m1") => format!("arch_e2m1_to_f32({b})"),
-                            Some("e2m3") => format!("arch_e2m3_to_f32({b})"),
-                            Some("e3m2") => format!("arch_e3m2_to_f32({b})"),
-                            Some("f32") => b,
-                            _ => {
-                                // int -> f32 (RNE) via the synthesizable helper.
-                                if self.expr_is_signed(base) {
-                                    format!("arch_i64_to_f32(64'($signed({b})))")
-                                } else {
-                                    format!("arch_u64_to_f32(64'($unsigned({b})))")
-                                }
-                            }
-                        }
-                    }
-                    "to_bf16" => {
-                        self.fp_helpers_used.set(true);
-                        match self.expr_float_fmt(base) {
-                            Some("f32") => format!("arch_f32_to_bf16({b})"),
-                            Some("bf16") => b,
-                            // fp8 -> bf16 is rejected by typecheck in v1
-                            // (route via .to_fp32()); keep the arm total.
-                            Some("e4m3") => format!("arch_f32_to_bf16(arch_e4m3_to_f32({b}))"),
-                            Some("e5m2") => format!("arch_f32_to_bf16(arch_e5m2_to_f32({b}))"),
-                            Some("e2m1") => format!("arch_f32_to_bf16(arch_e2m1_to_f32({b}))"),
-                            Some("e2m3") => format!("arch_f32_to_bf16(arch_e2m3_to_f32({b}))"),
-                            Some("e3m2") => format!("arch_f32_to_bf16(arch_e3m2_to_f32({b}))"),
-                            _ => {
-                                // int -> f32 (RNE) -> bf16 (RNE). DECLARED semantics
-                                // (issue #629, resolved as f32-routed / VR(f32)):
-                                // int.to_bf16() == narrow_bf16(f32(i)), matching the
-                                // bf16 fma f32-accumulate convention (PR #627). This
-                                // is a *double* rounding and is NOT correctly-rounded
-                                // int->bf16 for |i| >= 2^24 — the f32 step can land
-                                // exactly on a bf16 midpoint and tie-to-even the
-                                // wrong way (witness i=16842753 -> 0x4b80, correctly
-                                // rounded 0x4b81). Routing via f32 is hardware-
-                                // realistic (no direct int->bf16 in RISC-V) and
-                                // intended, not a bug — see doc/ARCH_HDL_Specification.md
-                                // §3.8 "Rounding convention" and doc/proposal_fp_rounding_semantics.md.
-                                // Sim mirror: src/sim_codegen _arch_{i,u}_to_bf16.
-                                if self.expr_is_signed(base) {
-                                    format!("arch_f32_to_bf16(arch_i64_to_f32(64'($signed({b}))))")
-                                } else {
-                                    format!(
-                                        "arch_f32_to_bf16(arch_u64_to_f32(64'($unsigned({b}))))"
-                                    )
-                                }
-                            }
-                        }
-                    }
-                    "to_fp8e4m3" | "to_fp8e5m2" | "to_fp4e2m1" | "to_fp6e2m3" | "to_fp6e3m2" => {
-                        self.fp_helpers_used.set(true);
-                        let (narrow, tgt) = match method.name.as_str() {
-                            "to_fp8e4m3" => ("arch_f32_to_e4m3", "e4m3"),
-                            "to_fp4e2m1" => ("arch_f32_to_e2m1", "e2m1"),
-                            "to_fp6e2m3" => ("arch_f32_to_e2m3", "e2m3"),
-                            "to_fp6e3m2" => ("arch_f32_to_e3m2", "e3m2"),
-                            _ => ("arch_f32_to_e5m2", "e5m2"),
-                        };
-                        match self.expr_float_fmt(base) {
-                            Some("f32") => format!("{narrow}({b})"),
-                            Some(t) if t == tgt => b,
-                            // BF16 / cross-fp8: exact widen, one narrow — CR.
-                            Some("bf16") => format!("{narrow}(arch_bf16_to_f32({b}))"),
-                            Some("e4m3") => format!("{narrow}(arch_e4m3_to_f32({b}))"),
-                            Some("e5m2") => format!("{narrow}(arch_e5m2_to_f32({b}))"),
-                            Some("e2m1") => format!("{narrow}(arch_e2m1_to_f32({b}))"),
-                            Some("e2m3") => format!("{narrow}(arch_e2m3_to_f32({b}))"),
-                            Some("e3m2") => format!("{narrow}(arch_e3m2_to_f32({b}))"),
-                            // Integers: int -> f32 (exact for the fp8-relevant
-                            // range, far below 2^24) -> one fp8 rounding — CR.
-                            _ => {
-                                if self.expr_is_signed(base) {
-                                    format!("{narrow}(arch_i64_to_f32(64'($signed({b}))))")
-                                } else {
-                                    format!("{narrow}(arch_u64_to_f32(64'($unsigned({b}))))")
-                                }
-                            }
-                        }
-                    }
-                    "to_uint" | "to_sint" => {
-                        self.fp_helpers_used.set(true);
-                        let width = args
-                            .first()
-                            .map(|w| self.emit_expr_str(w))
-                            .unwrap_or_else(|| "32".to_string());
-                        let wp = Self::paren_width(&width);
-                        let f32bits = match self.expr_float_fmt(base) {
-                            Some("bf16") => format!("arch_bf16_to_f32({b})"),
-                            Some("e4m3") => format!("arch_e4m3_to_f32({b})"),
-                            Some("e5m2") => format!("arch_e5m2_to_f32({b})"),
-                            _ => b.clone(),
-                        };
-                        // Toward-zero, saturating to the N-bit target, NaN -> type max
-                        // (synthesizable helper; returns a 64-bit value truncated to N).
-                        let conv = if method.name == "to_sint" {
-                            "arch_f32_to_sint"
-                        } else {
-                            "arch_f32_to_uint"
-                        };
-                        format!("{wp}'({conv}({f32bits}, {width}))")
-                    }
-                    _ => format!("{b}.{}()", method.name),
-                }
-            }
+            ExprKind::MethodCall(base, method, args) => self.emit_method_call_str(
+                base,
+                method,
+                args,
+                MethodCallHost::Main,
+                &|e: &Expr| self.emit_expr_str(e),
+                // The main emitter's chunked lowering re-emits the receiver
+                // itself; the pre-emitted receiver string is unused.
+                &|recv: &Expr, chunk: &Expr, _emitted: &str| {
+                    self.try_emit_reverse_chunked(recv, chunk)
+                },
+            ),
             ExprKind::Cast(expr, ty) => {
                 let e = self.emit_expr_str(expr);
                 match &**ty {
